reafctor: move int_hash_str to utils
